backup/requirements.txt
pillow
flask
werkzeug
